@@ -1,22 +1,25 @@
 {
   "name": "dotenv",
-  "version": "10.0.0",
+  "version": "16.0.1",
   "description": "Loads environment variables from .env file",
   "main": "lib/main.js",
+  "types": "lib/main.d.ts",
   "exports": {
-    ".": "./lib/main.js",
+    ".": {
+      "require": "./lib/main.js",
+      "types": "./lib/main.d.ts",
+      "default": "./lib/main.js"
+    },
     "./config": "./config.js",
     "./config.js": "./config.js",
     "./package.json": "./package.json"
   },
-  "types": "types/index.d.ts",
   "scripts": {
-    "flow": "flow",
-    "dtslint": "dtslint types",
+    "dts-check": "tsc --project tests/types/tsconfig.json",
     "lint": "standard",
-    "postlint": "standard-markdown",
-    "pretest": "npm run lint && npm run dtslint",
-    "test": "tap tests/*.js --100",
+    "lint-readme": "standard-markdown",
+    "pretest": "npm run lint && npm run dts-check",
+    "test": "tap tests/*.js --100 -Rspec",
     "prerelease": "npm test",
     "release": "standard-version"
   },
@@ -36,22 +39,18 @@
   "readmeFilename": "README.md",
   "license": "BSD-2-Clause",
   "devDependencies": {
-    "decache": "^4.5.1",
-    "dtslint": "^0.9.8",
-    "flow-bin": "^0.109.0",
-    "sinon": "^7.5.0",
-    "standard": "^13.1.0",
-    "standard-markdown": "^5.1.0",
-    "standard-version": "^7.0.0",
-    "tap": "^14.7.0"
+    "@types/node": "^17.0.9",
+    "decache": "^4.6.1",
+    "dtslint": "^3.7.0",
+    "sinon": "^12.0.1",
+    "standard": "^16.0.4",
+    "standard-markdown": "^7.1.0",
+    "standard-version": "^9.3.2",
+    "tap": "^15.1.6",
+    "tar": "^6.1.11",
+    "typescript": "^4.5.4"
   },
-  "dependencies": {},
   "engines": {
-    "node": ">=10"
-  },
-  "standard": {
-    "ignore": [
-      "flow-typed/"
-    ]
+    "node": ">=12"
   }
 }
